+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.0.0
 -----
 
+ * removed support for autowiring types
  * removed `Container::isFrozen`
  * removed support for dumping an ucompiled container in `PhpDumper`